# [TT-9949] do len check before compiling extendedPathSpec