<!-- Provide a general summary of your changes in the Title above -->

## Description
do len check before compiling extendedPathSpec

## Related Issue
https://tyktech.atlassian.net/browse/TT-9949

## Motivation and Context

<!-- Why is this change required? What problem does it solve? -->

## How This Has Been Tested
mprove getExtendedPathSpecs
Benchmark output

Before
```
go test -bench=Benchmark_getExtendedPathSpecs -run=^$ -benchtime=2s   -count 10
pkg: github.com/TykTechnologies/tyk/gateway
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_getExtendedPathSpecs-16        2344707              1011 ns/op               0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2462037               955.4 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2395380               950.3 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2599981               997.3 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2552166               955.4 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2362286              1089 ns/op               0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2038982              1193 ns/op               0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2137178              1009 ns/op               0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2344965               983.9 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16        2390276               993.9 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/TykTechnologies/tyk/gateway  35.058s
```

After 
```
go test -bench=Benchmark_getExtendedPathSpecs -run=^$ -benchtime=2s   -count 10
pkg: github.com/TykTechnologies/tyk/gateway
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_getExtendedPathSpecs-16       18663019               122.1 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       19080561               120.1 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       18848035               267.7 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       19545283               121.0 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       16387222               134.0 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       16352863               130.7 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       19724300               135.3 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       19792872               123.4 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       19063058               132.7 ns/op             0 B/op          0 allocs/op
Benchmark_getExtendedPathSpecs-16       16169594               126.7 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/TykTechnologies/tyk/gateway  33.713s
```
## Screenshots (if appropriate)

## Types of changes

<!-- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Refactoring or add test (improvements in base code or adds test coverage to functionality)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes that apply -->
<!-- If there are no documentation updates required, mark the item as checked. -->
<!-- Raise up any additional concerns not covered by the checklist. -->

- [ ] I ensured that the documentation is up to date
- [ ] I explained why this PR updates go.mod in detail with reasoning why it's required
- [ ] I would like a code coverage CI quality gate exception and have explained why