@Tyk-ITS 2023-09-01
**API tests result: success** :white_check_mark: 
Branch used: refs/pull/5505/merge
Commit: 84c9bc5ef5ceb974fa5dc0a93bdb2d94801b5be8 
Triggered by: pull_request (@jeffy-mathew)
[Execution page](https://github.com/TykTechnologies/tyk/actions/runs/6048607122)